+++ src/controllers/controller.js
@@ -11,10 +11,6 @@ const emojiFlags = require('emoji-flags')
 
 const { log } = require('../lib/logger')
 
-// TODO customize
-const defaultForms = [undefined, 163, 166, 169, 172, 175, 178, 181, 184, 187, 953, 956, 959, 616, 619, 622, 962, 965, 968, 45, 47, 971, 974, 697, 700, 598, 49, 51, 53, 776, 779, 782, 776, 785, 788, 981, 984, 55, 57, 987, 990, 157, 160, 265, 268, 271, 993, 996, 259, 262, 59, 61, 63, 65, 286, 289, 999, 1002, 280, 283, 235, 238, 241, 304, 307, 310, 809, 812, 815, 664, 667, 670, 1005, 1008, 67, 69, 71, 1011, 1014, 1017, 1020, 655, 658, 1023, 1026, 1029, 1032, 1035, 73, 75, 876, 879, 1038, 1041, 1044, 902, 214, 217, 870, 873, 1047, 1050, 729, 77, 224, 79, 713, 277, 1053, 703, 706, 846, 849, 1056, 1059, 839, 1062, 1065, 1068, 1071, 1074, 1077, 1080, 247, 1083, 640, 634, 898, 1086, 253, 256, 322, 1089, 1092, 1095, 1098, 1101, 677, 740, 743, 1104, 1107, 1110, 199, 716, 773, 836, 190, 193, 196, 135, 1115, 1118, 1121, 1124, 1127, 1130, 1133, 1136, 1139, 1142, 1145, 1148, 1151, 1154, 1157, 1160, 1163, 1166, 202, 1169, 1172, 1175, 1178, 1181, 1184, 1187, 1190, 1193, 646, 649, 652, 274, 1196, 1199, 1202, 244, 1205, 1208, 1211, 1214, 1217, 1220, 1223, 1226, 1229, 1232, 1235, 855, 1238, 719, undefined, 602, 1241, 1244, 1247, 1250, 803, 905, 1253, 1256, 1259, 250, 827, 1262, 797, 1265, 1268, 1271, 1274, 1277, 1280, 1283, 1286, 1289, 938, 1292, 1295, 229, 232, 1298, 1301, 1304, 680, 941, 1307, 1310, 1313, 1316, 1319, 1322, 1325, 1328, 1331, 1334, 1337, 313, 316, 319, 1340, 1343, 1346, 1349, 1352, 1355, 1358, 1361, 1364, 205, 208, 211, 1367, 1370, 945, 947, 600, 1379, 1382, 1385, 1388, 1391, 1394, 1397, 625, 628, 631, 1400, 1403, 1406, 1409, 292, 295, 298, 1412, 1415, 1418, 1421, 1424, 1427, 1430, 1433, 1436, 1439, 1442, 1445, 1448, 1451, 1454, 1457, 1460, 1463, 1466, 923, 833, 1469, 1472, 1475, 1478, 1481, 1484, 1487, 1490, 1493, 1496, 1499, 1502, 1505, 1508, 734, 737, 1511, 1514, 1517, 1520, 1523, 1526, 1529, undefined, 746, 749, 752, 610, 613, 1532, 1535, 1538, 1541, 1544, 1547, 1550, 1553, 1556, 1559, 1562, 1565, 1568, 1571, 1574, 1577, 1580, 1583, 29, 1586, 908, 911, 914, 917, 1589, 1592, 830, 1595, 926, 929, 1598, 1601, 1604, 1607, 1610, 1613, 1616, 1619, 755, 758, 761, 764, 767, 770, 1622, 1625, 1628, 1631, 1634, 1637, 1640, 1643, 1646, 33, 688, 691, 694, 818, 821, 824, 1649, 1652, 1655, 1658, 1661, 1664, 1667, 1670, 1673, 1676, 1679, 1682, 1685, 1688, 1691, 1694, 1697, 1700, 1703, undefined, undefined, 1712, 1715, 1718, 1721, 1724, 1727, 1730, undefined, undefined, undefined, 1742, 1745, 1748, 1751, 1754, 722, 858, 1757, 1760, 1763, 791, 794, 1766, 1769, 1772, 1775, 1778, 1781, 1784, 861, 864, 867, 1787, 1790, 1793, 888, 891, 1796, 1799, 1802, 1805, 1808, 1811, 1814, 1817, 932, 935, 800, 661, 1820, 852, 1823, 643, 637, 1826, 1829, 1832, 1835, 806, 1838, 683, 301, 1841, 920, 1844, 81, 1847, 1850, 1853, 1856, 1859, 1862, 1865, undefined, 1871, 1874, 1877, 1880, undefined, 100, 1886, 1889, 1892, 1895, 1898, 1901, 1904, 1907, 1910, 1913, 1916, 1919, 1922, 1925, 1928, 1931, 1934, 1937, 1940, 1943, 1946, 1949, 1952, 1955, 1958, 1961, 1964, 1967, 1970, 1973, 1976, 1979, 1982, 1985, 1988, 1991, 1994, 1997, 2000, 2003, 2006, 2009, 2012, 2015, 2018, 2021, 2024, 2027, 2030, 2033, 2036, 2039, 2042, 2045, 2048, 2051, undefined, 2054, 2057, 2060, 2063, undefined, 2066, 2069, 2072, 2075, 2078, 2081, 2084, 2087, 2090, 2093, 2096, 2099, 2102, 2105, 2108, 2111, 2114, 2117, 2120, 2123, 2126, 2129, 2132, 2135, 2138, 2141, 2144, 2147, 2150, undefined, undefined, 2153, 2156, 2159, 2162, 2165, 2168, 2171, 2174, 2177, 2180, 2183, 2186, 2189, 2192, 2195, 2198, 2201, 2204, 2207, 2210, 2213, 2216, 2219, 2222, 2225, 2228, 2231, 2234, 2237, 2240, 2243, 2246, 2249, 2252, 2255, 2258, 2261, 2264, 2267, 2270, 2273, 2276, 2279, 2282, 2285, 2288, 2291, 2294, 2297, 2300, 2303, 2306, 2309, 2312, undefined, undefined, 2315, 2318, undefined, 146, undefined, undefined, 593, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, 2321, 2324, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, 2501, 2504, undefined, 2510]
-const availablePokemon = ['598_2186', '025_598_female_12_shiny', '332_613_female_shiny', '351_30', '593_2171_female_shiny', '588_2156_shiny', '201_22_shiny', '458_1817_shiny', '104_224', '151_1115', '255_1358_shiny', '093_1041', '098_870', '482_1853_shiny', '220_1277', '050_59_shiny', '483_1856', '113_1056', '194_1226_female_shiny', '481_1850_shiny', '429_722_shiny', '366_1607_shiny', '453_1802_female', '026_49_07_shiny', '126_634_shiny', '114_1059', '004_172_14', '172_1175_04_shiny', '550_137_shiny', '638_2306', '371_755_shiny', '327_38', '308_1481_female', '459_932', '188_1208', '369_1616_female_shiny', '209_1253', '493_107_shiny', '026_50_04', '413_87_shiny', '457_1814', '596_2180_shiny', '606_2210_shiny', '106_713_shiny', '120_1074', '186_244_shiny', '479_83', '129_253_female', '322_1517', '376_770', '026_49_female_05_shiny', '289_1430', '609_2219_shiny', '167_1163_shiny', '025_598_female_10', '011_956_shiny', '201_27', '311_1490_shiny', '001_897', '057_1002_shiny', '310_1487_shiny', '321_1514', '123_247_female', '207_803', '004_172_11', '026_49_02', '089_76_shiny', '316_1505', '052_2335_shiny', '463_1820', '603_2201', '343_1562', '008_184_05_shiny', '314_1499', '615_2237_shiny', '201_8', '162_1148', '202_602_female_11', '182_274_shiny', '575_2123_shiny', '101_1050', '026_49_female', '327_121_shiny', '149_196', '009_187_05', '492_93_shiny', '375_767', '330_752', '621_2255', '342_1559_shiny', '037_56', '353_908', '542_2030_shiny', '207_803_female', '466_643_shiny', '020_48', '076_71', '124_1083_shiny', '415_1715_female_shiny', '321_1514_shiny', '648_151', '351_29_shiny', '593_2331_shiny', '572_2114', '175_1184_12', '242_1328_shiny', '272_1397_female_shiny', '488_1871_shiny', '585_587', '203_1241', '044_268_shiny', '025_894_shiny', '310_1487', '168_1166', '025_598_female_03', '074_67', '178_1193_female', '208_905_female', '025_2332_shiny', '553_2060', '459_932_female_shiny', '605_2207', '395_1655', '202_602_shiny', '005_175_shiny', '602_2198', '599_2189_shiny', '427_1751_12_shiny', '493_106', '470_1832_07_shiny', '204_1244_shiny', '465_1823_female_shiny', '404_1682', '643_2315_shiny', '391_821', '014_619_shiny', '413_89', '158_1136_shiny', '421_94_shiny', '517_1955_shiny', '154_1124_shiny', '368_1613', '501_1907', '493_111_shiny', '350_1583_female_shiny', '138_740_shiny', '265_600_shiny', '025_598_female_23', '410_1700', '536_2012', '615_2237', '026_49_09_shiny', '064_307', '537_2015', '231_1301_shiny', '461_800_female', '479_86_shiny', '025_598_03', '062_241', '303_833_shiny', '365_1604_shiny', '808_2321', '604_2204', '450_891_shiny', '025_598_female_04', '492_00', '378_1625_shiny', '083_2338', '574_2120', '637_2303', '484_1859', '354_911', '026_49_08', '552_2057', '534_2006', '254_1355', '090_876', '121_1077', '490_1877_shiny', '423_99', '365_1604', '217_1268_female_shiny', '036_984_shiny', '430_858', '419_1727_female_shiny', '070_667_shiny', '051_61', '186_244_female_shiny', '110_706', '050_60_shiny', '025_598_female_25_shiny', '089_75', '514_1946_shiny', '419_1727_shiny', '494_1886_shiny', '648_151_shiny', '061_238', '015_v1_shiny', '555_138_shiny', '360_1595_shiny', '563_2087', '028_53', '454_1805_shiny', '001_897_shiny', '323_1520_female_shiny', '572_2114_shiny', '172_1175_08_shiny', '020_47_female_11', '025_2332', '026_49', '030_779_shiny', '025_598_13_shiny', '192_1220', '338_1547_shiny', '210_1256_shiny', '058_280', '118_1068', '071_670', '401_1673', '467_637', '152_1118', '505_1919_shiny', '384_1643', '542_2030', '023_697', '166_1160_shiny', '493_117', '312_1493', '135_1098_07', '379_1628_shiny', '071_670_shiny', '048_259', '493_104_shiny', '428_1754', '077_2337_shiny', '202_602_female_11_shiny', '201_5_shiny', '451_1796', '020_47_female', '564_2090', '111_846_shiny', '437_1769', '306_1475', '418_1724_shiny', '228_229_shiny', '240_1322', '624_2264', '363_1598_shiny', '324_1523_shiny', '134_1095_shiny', '201_9_shiny', '592_2330_shiny', '454_1805', '019_46_shiny', '232_1304', '325_1526_shiny', '056_999', '025_598_05', '208_905_shiny', '026_49_female_05', '332_613_female', '066_809', '376_770_shiny', '563_2087_shiny', '344_1565_shiny', '513_1943', '143_199_shiny', '322_1517_female_shiny', '589_2159_shiny', '645_145_shiny', '585_586_shiny', '505_1919', '046_993_shiny', '646_146_shiny', '477_920_shiny', '053_65_shiny', '546_2042_shiny', '424_1742_female_shiny', '642_142_shiny', '201_11_shiny', '278_1406_shiny', '543_2033_shiny', '045_271_female_shiny', '329_749', '314_1499_shiny', '315_1502_female_shiny', '025_598_female_13_shiny', '077_2337', '177_1190_shiny', '386_00', '000', '180_649_shiny', '269_1388_female_shiny', '396_1658_female_shiny', '317_1508_shiny', '487_00_shiny', '257_1364_female', '191_1217_shiny', '037_56_shiny', '091_879', '136_1101_07_shiny', '465_1823_shiny', '521_1967_shiny', '420_1730', '107_277', '386_34', '500_1904', '648_152_shiny', '096_214_shiny', '460_935', '127_898', '322_1517_shiny', '316_1505_shiny', '445_867_female_shiny', '551_2054_shiny', '249_1340_shiny', '393_1649', '620_2252_shiny', '415_1715', '172_1175_04', '421_00', '068_815', '601_2195_shiny', '487_91_shiny', '478_1844', '649_595', '525_1979', '183_1196', '213_827', '025_598_12_shiny', '554_2063_shiny', '076_72', '137_677', '529_1991_shiny', '316_1505_female_shiny', '084_1026_female', '077_2336', '318_734_shiny', '592_2168_female', '069_664', '410_1700_shiny', '323_1520', '641_140', '201_13_shiny', '442_1784', '345_1568', '363_1598', '025_598_12', '571_2111', '407_1691_female', '058_280_shiny', '136_1101_shiny', '292_1439', '118_1068_female_shiny', '493_114', '649_594', '007_895', '493_102', '446_1787_shiny', '336_1541', '094_1044', '178_1193_shiny', '026_49_female_07_shiny', '308_1481_shiny', '198_855_female_shiny', '130_256', '461_800', '185_1202', '809_2324_shiny', '201_3_shiny', '172_1175_02', '112_849', '418_1724', '295_1448', '270_1391_shiny', '284_1415', '599_2189', '608_2216', '195_1229', '234_941_shiny', '421_95', '635_2297', '316_1505_female', '077_1011', '026_50_02', '269_1388', '189_1211_shiny', '612_2228', '587_2153', '625_2267', '622_2258', '493_113_shiny', '129_253_female_shiny', '317_1508_female_shiny', '042_160_female_shiny', '350_1583_female', '460_935_female_shiny', '493_114_shiny', '038_57_shiny', '350_1583', '327_37_shiny', '025_901', '447_1790_shiny', '493_109', '025_598_13', '203_1241_shiny', '622_2258_shiny', '456_1811_female_shiny', '043_265', '380_1631', '561_2081', '540_2024_shiny', '006_v2_shiny', '418_1724_female_shiny', '416_1718_shiny', '039_987_shiny', '620_2252', '009_187_05_shiny', '143_199', '623_2261', '339_1550_shiny', '147_190', '009_v1', '045_271_shiny', '201_22', '062_241_shiny', '474_683', '369_1616_female', '319_737_shiny', '643_2315', '267_1382_shiny', '511_1937_shiny', '361_926_shiny', '390_818_shiny', '102_729', '493_102_shiny', '009_952', '034_788_shiny', '476_1841', '600_2192_shiny', '396_1658', '311_1490', '234_941_08_shiny', '132_1089', '446_1787', '248_319', '435_794_shiny', '633_2291', '329_749_shiny', '263_945_shiny', '215_797', '247_316_shiny', '553_2060_shiny', '063_304_shiny', '025_598_23_shiny', '568_2102', '097_217', '169_202', '607_2213', '286_1421', '098_870_shiny', '201_21_shiny', '266_1379', '276_1400_shiny', '214_1262_shiny', '123_247_shiny', '088_74_shiny', '649_593', '649_596_shiny', '256_1361_female', '213_827_shiny', '095_902_shiny', '649_595_shiny', '585_585_shiny', '862_2501', '227_1295_shiny', '647_150', '216_1265', '389_694', '025_894', '453_1802', '560_2078', '148_193_shiny', '133_1092_shiny', '348_1577', '487_90', '427_1751_shiny', '479_81', '424_1742', '456_1811_shiny', '014_619', '569_2105_shiny', '224_1289', '492_93', '422_00_shiny', '602_2198_shiny', '195_1229_female_shiny', '647_149_shiny', '194_1226_female', '109_703_shiny', '348_1577_shiny', '632_2288', '386_35_shiny', '201_10_shiny', '608_2216_shiny', '003_169_shiny', '422_97', '197_1235_07', '059_283_shiny', '035_981', '440_1778_shiny', '541_2027_shiny', '471_1835_shiny', '327_37', '649_597_shiny', '614_2234', '645_144_shiny', '166_1160', '416_1718', '480_1847_shiny', '434_791_shiny', '025_598_11', '201_00', '175_1184_12_shiny', '479_00_shiny', '450_891_female', '212_250_female', '050_60', '493_108_shiny', '567_2099_shiny', '221_1280_female', '267_1382_female', '138_740', '001_163_14', '280_292', '473_1838_female', '508_1928', '052_64_shiny', '142_1110_shiny', '351_00', '001_163', '291_1436', '642_142', '122_1080_shiny', '313_1496_shiny', '051_61_shiny', '190_1214', '006_951', '108_1053', '079_1017_shiny', '041_157_female', '448_1793_shiny', '116_1062', '026_49_female_09_shiny', '162_1148_shiny', '519_1961_shiny', '481_1850', '555_2343_shiny', '302_923', '225_938_shiny', '388_691_shiny', '154_1124_female_shiny', '045_271_female', '359_830', '610_2222_shiny', '312_1493_shiny', '417_1721_female', '649_596', '630_2282', '333_1532', '127_898_shiny', '025_598_24', '570_2108_shiny', '417_1721_female_shiny', '201_15_shiny', '085_1029_female', '559_2075', '272_1397_female', '026_49_female_01', '149_196_shiny', '387_688', '510_1934', '025_598_female_03_shiny', '109_703', '089_75_shiny', '026_49_03', '258_205_shiny', '026_50_04_shiny', '026_50_01_shiny', '465_1823', '111_846', '496_1892', '528_1988', '594_2174', '052_2335', '019_45_female_shiny', '395_1655_shiny', '299_1460', '003_950_female', '240_1322_shiny', '072_1005_shiny', '051_62_shiny', '084_1026_female_shiny', '047_996', '092_1038', '006_v3', '238_1316', '105_80_shiny', '370_1619_shiny', '222_1283_shiny', '124_1083', '201_24_shiny', '439_1775', '111_846_female_shiny', '226_1292_shiny', '426_1748', '289_1430_shiny', '412_118_shiny', '181_652_shiny', '618_2345_shiny', '454_1805_female_shiny', '372_758', '327_121', '025_598_02_shiny', '147_190_shiny', '201_10', '019_45_female', '459_932_female', '025_598_female_12', '633_2291_shiny', '623_2261_shiny', '096_214', '614_2234_shiny', '264_948', '092_1038_shiny', '026_49_09', '251_1346', '413_00_shiny', '487_91', '026_49_female_03', '415_1715_female', '417_1721', '218_1271_shiny', '511_1937', '479_82', '262_1370', '588_2156', '327_44', '159_1139_shiny', '386_00_shiny', '094_1044_11_shiny', '454_1805_female', '464_852_shiny', '391_821_shiny', '163_1151_shiny', '433_1763_shiny', '315_1502', '808_2321_shiny', '117_1065', '082_658', '419_1727_female', '576_2126_shiny', '261_1367', '626_2270_shiny', '486_1865', '026_49_08_shiny', '026_49_04_shiny', '170_1169', '083_1023', '470_1832_07', '025_598_female_22_shiny', '592_2168_female_shiny', '175_1184_shiny', '201_2_shiny', '004_172_shiny', '023_697_shiny', '485_1862', '239_1319', '414_1712', '461_800_female_shiny', '339_1550', '025_598_female_24_shiny', '490_1877', '548_2048_shiny', '265_600', '214_1262_female', '026_49_female_07', '232_1304_female_shiny', '290_1433', '032_776_shiny', '501_1907_shiny', '182_274', '462_661', '413_89_shiny', '007_181_05_shiny', '379_1628', '166_1160_female', '607_2213_shiny', '123_247', '423_99_shiny', '192_1220_shiny', '141_1107_shiny', '557_2069', '532_2000', '384_1643_shiny', '282_298_shiny', '074_68_shiny', '165_1157_female', '301_1466', '130_256_female', '264_947_shiny', '234_941_08', '012_959', '351_29', '397_1661', '008_184_shiny', '025_598_shiny', '478_1844_shiny', '025_598_female_10_shiny', '186_244', '112_849_female', '025_598_08', '364_1601', '449_888_female', '068_815_shiny', '863_2504', '024_700_shiny', '450_891', '539_2021', '224_1289_female_shiny', '015_622', '201_19', '493_112', '025_2332_female_shiny', '640_2312_shiny', '516_1952_shiny', '172_1175_05', '360_1595', '101_1050_shiny', '004_896', '157_1133', '033_785_shiny', '382_1637', '064_307_female_shiny', '537_2015_shiny', '592_2168', '469_1829', '580_2138_shiny', '201_4', '203_1241_female_shiny', '294_1445', '422_96', '025_598_female_01_shiny', '335_1538_shiny', '637_2303_shiny', '201_1_shiny', '336_1541_shiny', '245_1337', '025_598_female_11', '519_1961', '493_00', '293_1442_shiny', '555_2342_shiny', '019_45_shiny', '286_1421_shiny', '383_1640', '390_818', '002_166_shiny', '189_1211', '001_163_shiny', '219_1274', '518_1958_shiny', '050_59', '253_1352', '636_2300_shiny', '159_1139', '020_47', '593_2171_female', '400_1670_shiny', '413_00', '201_9', '422_00', '528_1988_shiny', '170_1169_shiny', '105_79', '561_2081_shiny', '160_1142_shiny', '307_1478', '202_602', '257_1364_shiny', '521_1967_female_shiny', '026_50_01', '582_2144_shiny', '026_49_female_08', '487_90_shiny', '128_1086', '521_1967_female', '646_146', '184_1199_shiny', '252_1349_shiny', '421_95_shiny', '351_31', '243_1331', '201_6', '493_117_shiny', '280_292_shiny', '063_304', '094_1044_shiny', '177_1190', '346_1571', '190_1214_female_shiny', '201_25', '060_235', '055_289', '075_70_shiny', '551_2054', '028_53_shiny', '278_1406', '026_49_female_04_shiny', '424_1742_female', '272_1397_shiny', '577_2129', '006_v3_shiny', '362_929', '048_259_shiny', '447_1790', '473_1838_shiny', '202_602_female_shiny', '455_1808', '012_959_female', '015_622_shiny', '075_69_shiny', '115_839', '517_1955', '340_1553_shiny', '077_2583_shiny', '206_1250', '025_598_female_25', '485_1862_shiny', '277_1403_shiny', '151_1115_shiny', '315_1502_shiny', '495_1889_shiny', '020_47_11', '126_634', '160_1142', '026_50', '010_953', '056_999_shiny', '052_63', '091_879_shiny', '110_944', '203_1241_female', '020_48_shiny', '578_2132_shiny', '891_00_shiny', '489_1874', '524_1976_shiny', '166_1160_female_shiny', '102_729_shiny', '224_1289_female', '007_181_14', '264_947', '025_598_25_shiny', '333_1532_shiny', '468_1826_shiny', '026_50_02_shiny', '544_2036_shiny', '079_1017', '090_876_shiny', '298_1457_shiny', '531_1997', '491_1880', '074_68', '025_894_female_shiny', '049_262', '026_49_04', '464_852', '493_110_shiny', '484_1859_shiny', '201_6_shiny', '368_1613_shiny', '324_1523', '509_1931', '087_1035', '493_104', '592_2330', '233_680_shiny', '026_49_female_03_shiny', '172_1175_07', '644_2318', '515_1949', '196_1232_07', '120_1074_shiny', '335_1538', '381_1634_shiny', '267_1382_female_shiny', '493_108', '026_49_03_shiny', '119_1071_female', '116_1062_shiny', '026_49_06_shiny', '290_1433_shiny', '613_2231', '201_24', '489_1874_shiny', '025_598_female_24', '026_49_female_09', '025_894_female', '427_1751', '137_677_shiny', '025_598_female_02', '065_310', '025_598_02', '562_2084_shiny', '026_49_female_08_shiny', '523_1973', '257_1364_female_shiny', '493_112_shiny', '172_1175_03', '892_00_shiny', '204_1244', '201_23', '103_77', '423_98_shiny', '251_1346_shiny', '025_598_22', '327_41', '459_932_shiny', '201_20', '118_1068_female', '575_2123', '611_2225_shiny', '407_1691_shiny', '287_1424', '025_598_female_09_shiny', '106_713', '535_2009_shiny', '277_1403', '033_785_11_shiny', '456_1811', '198_855_female', '402_1676_female_shiny', '197_1235_shiny', '191_1217', '221_1280_female_shiny', '081_655', '155_1127_shiny', '059_283', '389_694_shiny', '423_98', '558_2072', '225_938', '025_598_10_shiny', '207_803_shiny', '241_1325_shiny', '579_2135_shiny', '367_1610', '256_1361_female_shiny', '565_2093', '044_268', '586_592', '449_888', '026_49_female_shiny', '513_1943_shiny', '498_1898', '586_591_shiny', '530_1994_shiny', '007_181_11', '355_914', '041_157_female_shiny', '027_51', '313_1496', '224_1289_shiny', '007_181_05', '477_920', '370_1619', '432_1760_shiny', '590_2162_shiny', '025_598_01_shiny', '172_1175_01_shiny', '172_1175_05_shiny', '202_602_11_shiny', '328_746', '154_1124', '493_110', '234_941', '495_1889', '163_1151', '493_109_shiny', '423_00_shiny', '206_1250_shiny', '201_11', '499_1901_shiny', '403_1679_female', '009_v1_shiny', '075_69', '002_166', '493_115', '197_1235_07_shiny', '463_1820_shiny', '024_700', '549_2051', '327_40', '398_1664_female', '374_764', '236_1310_shiny', '407_1691', '555_2342', '493_111', '392_824', '025_598_female_06', '272_1397', '222_1283', '041_157', '164_1154', '619_2249_shiny', '259_208_shiny', '185_1202_female_shiny', '291_1436_shiny', '354_911_shiny', '606_2210', '176_1187', '345_1568_shiny', '076_71_shiny', '512_1940_shiny', '038_57', '492_92', '284_1415_shiny', '377_1622', '386_36_shiny', '455_1808_shiny', '378_1625', '327_41_shiny', '309_1484_shiny', '103_78_shiny', '025_598_female', '067_812_shiny', '567_2099', '586_590', '172_1175_08', '069_664_shiny', '201_7', '025_598_female_04_shiny', '173_1178_shiny', '398_1664', '584_2150', '031_782', '190_1214_female', '559_2075_shiny', '597_2183', '205_1247_shiny', '026_50_03', '217_1268_female', '355_914_shiny', '341_1556_shiny', '283_1412', '210_1256', '529_1991', '125_640', '394_1652', '585_586', '624_2264_shiny', '026_49_female_06', '479_83_shiny', '012_959_female_shiny', '403_1679', '273_625', '156_1130_shiny', '230_1298', '421_00_shiny', '364_1601_shiny', '190_1214_shiny', '047_996_shiny', '288_1427', '619_2249', '016_962', '150_135_shiny', '025_949_female', '025_598_09_shiny', '414_1712_shiny', '418_1724_female', '299_1460_shiny', '326_1529', '283_1412_shiny', '025_598_female_23_shiny', '399_1667_shiny', '025_598_11_shiny', '327_00_shiny', '344_1565', '025_598_female_09', '025_598_female_13', '133_1092_07_shiny', '256_1361', '201_12', '432_1760', '367_1610_shiny', '165_1157_female_shiny', '161_1145', '482_1853', '628_2276', '641_141_shiny', '003_950', '051_62', '236_1310', '585_588', '589_2159', '340_1553', '169_202_shiny', '296_1451', '425_1745', '862_2501_shiny', '285_1418_shiny', '081_655_shiny', '025_598_female_08', '134_1095', '356_917_shiny', '003_169_female_shiny', '198_855', '257_1364', '441_1781_shiny', '403_1679_shiny', '136_1101_07', '026_50_03_shiny', '543_2033', '493_105_shiny', '136_1101', '217_1268', '479_84', '470_1832_shiny', '238_1316_shiny', '153_1121_shiny', '035_981_shiny', '582_2144', '483_1856_shiny', '403_1679_female_shiny', '201_23_shiny', '407_1691_female_shiny', '248_319_shiny', '444_864_female', '473_1838', '049_262_shiny', '472_806', '431_1757', '629_2279', '250_1343', '020_47_female_shiny', '892_00', '130_256_shiny', '533_2003', '544_2036', '131_322', '212_250_female_shiny', '486_1865_shiny', '067_812', '029_776_shiny', '201_25_shiny', '253_1352_shiny', '001_163_14_shiny', '025_598_22_shiny', '104_224_shiny', '374_764_shiny', '201_26_shiny', '060_235_shiny', '216_1265_shiny', '337_1544', '493_116', '550_136_shiny', '625_2267_shiny', '497_1895', '025_598_female_05_shiny', '131_322_shiny', '208_905', '274_628_female_shiny', '031_782_shiny', '554_2341', '550_137', '439_1775_shiny', '357_1589', '547_2045', '184_1199', '307_1478_female', '565_2093_shiny', '070_667', '119_1071_shiny', '008_184_05', '520_1964', '005_175', '266_1379_shiny', '201_26', '500_1904_shiny', '171_1172', '471_1835_07', '555_139_shiny', '320_1511_shiny', '571_2111_shiny', '139_743_shiny', '380_1631_shiny', '025_598_female_02_shiny', '399_1667_female_shiny', '200_719_shiny', '564_2090_shiny', '351_30_shiny', '042_160', '401_1673_female', '522_1970', '460_935_female', '113_1056_shiny', '435_794', '245_1337_shiny', '075_70', '462_661_shiny', '423_00', '223_1286', '201_13', '006_178', '097_217_female_shiny', '083_1023_shiny', '144_716', '425_1745_shiny', '185_1202_shiny', '605_2207_shiny', '453_1802_shiny', '398_1664_shiny', '201_14_shiny', '438_1772_shiny', '100_1047', '085_1029_female_shiny', '201_27_shiny', '404_1682_shiny', '635_2297_shiny', '034_788', '146_836', '397_1661_female_shiny', '187_1205_shiny', '201_14', '027_52_shiny', '645_144', '649_593_shiny', '555_138', '094_1044_11', '503_1913', '195_1229_shiny', '570_2108', '193_1223_shiny', '550_136', '469_1829_shiny', '566_2096_shiny', '639_2309', '065_310_female_shiny', '584_2150_shiny', '317_1508', '326_1529_shiny', '004_172', '458_1817', '387_688_shiny', '603_2201_shiny', '647_150_shiny', '112_849_female_shiny', '207_803_female_shiny', '043_265_shiny', '053_66_shiny', '218_1271', '327_38_shiny', '351_32_shiny', '865_2510', '130_256_female_shiny', '628_2276_shiny', '406_1688_shiny', '020_47_shiny', '554_2063', '018_v1', '430_858_shiny', '327_42_shiny', '508_1928_shiny', '591_2165_shiny', '600_2192', '029_776', '208_905_female_shiny', '516_1952', '196_1232_shiny', '397_1661_shiny', '026_49_female_04', '480_1847', '199_1238', '371_755', '298_1457', '593_2171', '025_598_01', '114_1059_shiny', '352_1586_shiny', '305_1472', '492_00_shiny', '150_133_shiny', '201_17_shiny', '158_1136', '359_830_shiny', '522_1970_shiny', '082_658_shiny', '039_987', '585_588_shiny', '025_598_female_08_shiny', '396_1658_shiny', '148_193', '412_120_shiny', '586_592_shiny', '493_107', '243_1331_shiny', '025_901_shiny', '497_1895_shiny', '099_873', '583_2147_shiny', '627_2273', '026_49_01', '327_00', '295_1448_shiny', '007_181', '436_1766', '631_2285', '037_55_shiny', '194_1226_shiny', '178_1193', '351_31_shiny', '366_1607', '242_1328', '539_2021_shiny', '394_1652_shiny', '453_1802_female_shiny', '634_2294_shiny', '593_2171_shiny', '592_2168_shiny', '201_7_shiny', '172_1175_shiny', '293_1442', '308_1481_female_shiny', '472_806_shiny', '479_82_shiny', '201_00_shiny', '255_1358_female', '172_1175_07_shiny', '021_971_shiny', '342_1559', '155_1127', '249_1340', '421_94', '055_289_shiny', '865_2510_shiny', '468_1826', '117_1065_shiny', '547_2045_shiny', '304_1469', '275_631_female_shiny', '471_1835_07_shiny', '110_944_shiny', '593_2331', '053_66', '057_1002', '448_1793', '405_1685_shiny', '282_298', '466_643', '556_2066_shiny', '327_42', '351_32', '331_610', '327_43', '411_1703_shiny', '201_28_shiny', '412_118', '027_52', '475_301_shiny', '356_917', '479_85_shiny', '025_598_female_07', '362_929_shiny', '201_4_shiny', '146_836_shiny', '003_v1_shiny', '451_1796_shiny', '018_968', '018_v1_shiny', '581_2141', '457_1814_female_shiny', '084_1026_shiny', '276_1400', '201_28', '129_253_shiny', '012_959_shiny', '424_1742_shiny', '617_2243_shiny', '152_1118_shiny', '487_00', '105_79_shiny', '186_244_female', '627_2273_shiny', '004_172_14_shiny', '052_64', '271_1394_shiny', '597_2183_shiny', '332_613', '172_1175_01', '270_1391', '585_587_shiny', '612_2228_shiny', '456_1811_female', '275_631_shiny', '479_84_shiny', '386_33', '165_1157', '450_891_female_shiny', '077_2583', '026_50_05_shiny', '433_1763', '093_1041_shiny', '007_895_shiny', '437_1769_shiny', '465_1823_female', '088_73_shiny', '294_1445_shiny', '201_19_shiny', '025_598_06', '110_706_shiny', '172_1175_12', '025_949', '040_990', '398_1664_female_shiny', '504_1916', '555_2343', '211_1259', '215_797_shiny', '319_737', '100_1047_shiny', '275_631', '185_1202_female', '201_8_shiny', '429_722', '135_1098_shiny', '227_1295', '201_18_shiny', '616_2240', '343_1562_shiny', '385_1646_shiny', '025_598_06_shiny', '618_2246_shiny', '404_1682_female', '809_2324', '139_743', '512_1940', '353_908_shiny', '025_598_09', '400_1670_female_shiny', '078_1014_shiny', '549_2051_shiny', '006_178_shiny', '103_77_shiny', '586_591', '076_72_shiny', '003_v1', '595_2177', '053_65', '260_211', '141_1107', '229_232', '445_867_female', '025_598_08_shiny', '524_1976', '461_800_shiny', '017_965', '042_160_shiny', '323_1520_shiny', '228_229', '026_49_female_01_shiny', '019_45', '475_301', '532_2000_shiny', '493_105', '510_1934_shiny', '036_984', '013_616_shiny', '586_589', '250_1343_shiny', '636_2300', '171_1172_shiny', '026_49_07', '020_47_female_11_shiny', '545_2039_shiny', '493_113', '088_73', '167_1163', '409_1697', '027_51_shiny', '054_286_shiny', '533_2003_shiny', '042_160_female', '548_2048', '443_861', '406_1688', '649_597', '041_157_shiny', '025_598_female_07_shiny', '473_1838_female_shiny', '065_310_female', '087_1035_shiny', '025_598', '201_16_shiny', '119_1071_female_shiny', '044_268_female_shiny', '452_1799', '196_1232', '025_598_23', '022_974', '021_971', '499_1901', '443_861_female', '097_217_female', '386_34_shiny', '347_1574_shiny', '025_598_female_22', '064_307_shiny', '061_238_shiny', '064_307_female', '560_2078_shiny', '168_1166_shiny', '493_00_shiny', '449_888_female_shiny', '639_2309_shiny', '642_143', '153_1121', '212_250', '306_1475_shiny', '498_1898_shiny', '003_169', '077_1011_shiny', '442_1784_shiny', '507_1925_shiny', '573_2117_shiny', '297_1454_shiny', '530_1994', '073_1008_shiny', '201_18', '297_1454', '372_758_shiny', '426_1748_shiny', '307_1478_female_shiny', '161_1145_shiny', '201_3', '630_2282_shiny', '420_1730_shiny', '493_101_shiny', '393_1649_shiny', '199_1238_shiny', '327_43_shiny', '252_1349', '099_873_shiny', '111_846_female', '415_1715_shiny', '221_1280_shiny', '174_1181_shiny', '172_1175_03_shiny', '646_147', '028_54', '577_2129_shiny', '135_1098', '648_152', '479_00', '132_1089_shiny', '401_1673_shiny', '025_598_female_shiny', '626_2270', '118_1068_shiny', '239_1319_shiny', '338_1547', '417_1721_shiny', '038_58_shiny', '304_1469_shiny', '540_2024', '470_1832', '172_1175_02_shiny', '229_232_female_shiny', '536_2012_shiny', '237_1313_shiny', '133_1092_11', '538_2018', '180_649', '518_1958', '254_1355_shiny', '015_v1', '097_217_shiny', '347_1574', '133_1092_07', '026_50_shiny', '305_1472_shiny', '275_631_female', '244_1334_shiny', '054_286', '285_1418', '108_1053_shiny', '175_1184', '604_2204_shiny', '863_2504_shiny', '080_1020', '073_1008', '026_49_01_shiny', '134_1095_07_shiny', '045_271', '434_791', '229_232_female', '201_16', '646_148', '645_145', '386_35', '281_295_shiny', '268_1385', '526_1982_shiny', '520_1964_shiny', '150_135', '246_313_shiny', '400_1670', '621_2255_shiny', '025_598_female_11_shiny', '493_116_shiny', '263_945', '074_67_shiny', '488_1871', '010_953_shiny', '233_680', '308_1481', '105_80', '033_785', '135_1098_07_shiny', '279_1409_shiny', '580_2138', '385_1646', '373_761_shiny', '201_5', '231_1301', '115_839_shiny', '273_625_shiny', '187_1205', '197_1235', '493_103_shiny', '422_96_shiny', '386_36', '392_824_shiny', '173_1178', '007_181_shiny', '007_181_11_shiny', '609_2219', '202_602_female', '154_1124_female', '237_1313', '558_2072_shiny', '219_1274_shiny', '303_833', '292_1439_shiny', '325_1526', '133_1092_11_shiny', '555_139', '646_148_shiny', '341_1556', '471_1835', '445_867_shiny', '017_965_shiny', '281_295', '616_2240_shiny', '174_1181', '315_1502_female', '212_250_shiny', '020_47_11_shiny', '026_49_female_06_shiny', '145_773_shiny', '214_1262', '598_2186_shiny', '129_253', '618_2246', '502_1910_shiny', '617_2243', '322_1517_female', '402_1676_female', '405_1685_female_shiny', '201_21', '350_1583_shiny', '568_2102_shiny', '641_140_shiny', '011_956', '025_598_07_shiny', '531_1997_shiny', '586_590_shiny', '443_861_female_shiny', '301_1466_shiny', '028_54_shiny', '009_187', '545_2039', '112_849_shiny', '209_1253_shiny', '025_598_04', '268_1385_shiny', '025_598_female_06_shiny', '164_1154_shiny', '327_44_shiny', '525_1979_shiny', '080_1020_shiny', '052_63_shiny', '016_962_shiny', '566_2096', '383_1640_shiny', '330_752_shiny', '066_809_shiny', '402_1676', '019_46', '337_1544_shiny', '552_2057_shiny', '008_184', '287_1424_shiny', '119_1071', '601_2195', '255_1358_female_shiny', '265_600_11', '026_49_05', '647_149', '578_2132', '262_1370_shiny', '300_1463', '142_1110', '025_598_07', '007_181_14_shiny', '460_935_shiny', '133_1092', '046_993', '065_310_shiny', '506_1922_shiny', '373_761', '491_1880_shiny', '523_1973_shiny', '009_187_shiny', '574_2120_shiny', '088_74', '181_652', '172_1175_12_shiny', '037_55', '085_1029', '026_49_female_02_shiny', '583_2147', '581_2141_shiny', '003_169_female', '441_1781', '327_39', '201_1', '610_2222', '638_2306_shiny', '479_85', '349_1580', '493_106_shiny', '264_948_shiny', '033_785_11', '269_1388_female', '274_628', '125_640_shiny', '004_896_shiny', '274_628_female', '534_2006_shiny', '573_2117', '375_767_shiny', '399_1667', '001_163_11', '255_1358', '579_2135', '201_17', '631_2285_shiny', '613_2231_shiny', '211_1259_shiny', '084_1026', '263_946', '232_1304_female', '546_2042', '121_1077_shiny', '509_1931_shiny', '296_1451_shiny', '408_1694', '172_1175', '026_49_female_02', '535_2009', '128_1086_shiny', '382_1637_shiny', '412_119', '388_691', '201_12_shiny', '507_1925', '032_776', '504_1916_shiny', '244_1334', '300_1463_shiny', '436_1766_shiny', '428_1754_shiny', '188_1208_shiny', '331_610_shiny', '449_888_shiny', '467_637_shiny', '263_946_shiny', '030_779', '457_1814_shiny', '349_1580_shiny', '220_1277_shiny', '145_773', '642_143_shiny', '077_2336_shiny', '358_1592', '554_2341_shiny', '085_1029_shiny', '369_1616_shiny', '479_86', '412_119_shiny', '038_58', '514_1946', '150_133', '317_1508_female', '302_923_shiny', '232_1304_shiny', '179_646_shiny', '440_1778', '644_2318_shiny', '586_589_shiny', '179_646', '202_602_11', '318_734', '122_1080', '408_1694_shiny', '474_683_shiny', '200_719', '618_2345', '230_1298_shiny', '585_585', '215_797_female', '089_76', '013_616', '641_141', '352_1586', '267_1382', '022_974_shiny', '205_1247', '556_2066', '323_1520_female', '493_103', '025_598_04_shiny', '413_87', '004_172_11_shiny', '107_277_shiny', '103_78', '413_88', '891_00', '258_205', '521_1967', '422_97_shiny', '444_864', '493_101', '332_613_shiny', '260_211_shiny', '178_1193_female_shiny', '086_1032', '026_49_05_shiny', '025_598_24_shiny', '400_1670_female', '235_1307_shiny', '405_1685_female', '503_1913_shiny', '346_1571_shiny', '072_1005', '025_598_05_shiny', '632_2288_shiny', '457_1814_female', '271_1394', '357_1589_shiny', '386_33_shiny', '083_2338_shiny', '261_1367_shiny', '134_1095_07', '404_1682_female_shiny', '412_120', '274_628_shiny', '095_902', '476_1841_shiny', '221_1280', '351_00_shiny', '494_1886', '086_1032_shiny', '527_1985_shiny', '576_2126', '595_2177_shiny', '444_864_shiny', '040_990_shiny', '527_1985', '479_81_shiny', '214_1262_female_shiny', '538_2018_shiny', '144_716_shiny', '044_268_female', '025_598_female_05', '025_598_25', '396_1658_female', '444_864_female_shiny', '217_1268_shiny', '629_2279_shiny', '256_1361_shiny', '026_50_05', '409_1697_shiny', '399_1667_female', '405_1685', '541_2027', '381_1634', '259_208', '445_867', '269_1388_shiny', '307_1478_shiny', '198_855_shiny', '502_1910', '328_746_shiny', '156_1130', '464_852_female_shiny', '327_39_shiny', '006_v2', '157_1133_shiny', '025_598_10', '402_1676_shiny', '419_1727', '438_1772', '493_115_shiny', '594_2174_shiny', '229_232_shiny', '334_1535_shiny', '195_1229_female', '334_1535', '427_1751_12', '496_1892_shiny', '026_49_02_shiny', '377_1622_shiny', '288_1427_shiny', '196_1232_07_shiny', '591_2165', '201_2', '223_1286_shiny', '201_20_shiny', '358_1592_shiny', '562_2084', '411_1703', '431_1757_shiny', '176_1187_shiny', '640_2312', '246_313', '183_1196_shiny', '235_1307', '018_968_shiny', '165_1157_shiny', '078_1014', '569_2105', '369_1616', '506_1922', '557_2069_shiny', '215_797_female_shiny', '320_1511', '515_1949_shiny', '194_1226', '140_1104_shiny', '026_49_06', '587_2153_shiny', '241_1325', '026_49_shiny', '413_88_shiny', '025_2332_female', '452_1799_shiny', '464_852_female', '649_594_shiny', '201_15', '443_861_shiny', '611_2225', '397_1661_female', '361_926', '401_1673_female_shiny', '526_1982', '596_2180', '247_316', '265_600_11_shiny', '123_247_female_shiny', '646_147_shiny', '634_2294', '140_1104', '327_40_shiny', '590_2162', '001_163_11_shiny', '279_1409', '309_1484', '025_598_female_01', '492_92_shiny', '193_1223', '226_1292', '025_598_03_shiny']
-
 class Controller {
 	constructor(db, config, dts, geofence, monsterData, discordCache, translator, mustache, weatherController) {
 		this.db = db
@@ -31,49 +27,6 @@ class Controller {
 		this.earthRadius = 6371 * 1000 // m
 		this.weatherController = weatherController
 		this.controllerData = {}
-	}
-
-	resolvePokemonIcon(pokemonId, pokemonForm = 0, evolutionId = 0, female = false, costumeId = 0, shiny = false) {
-		const pokeId = pokemonId.padStart(3, '0')
-		const costumeSuffix = costumeId ? `_${costumeId}` : ''
-		const shinySuffix = shiny ? '_shiny' : ''
-		if (evolutionId) {
-			let pokemonIdString = `${pokeId}_v${evolutionId}${costumeSuffix}${shinySuffix}`
-			if (availablePokemon.has(pokemonIdString)) return pokemonIdString
-			pokemonIdString = `${pokeId}_v${evolutionId}${shinySuffix}`
-			if (availablePokemon.has(pokemonIdString)) return pokemonIdString
-		}
-		const femaleSuffix = female ? '_female' : ''
-		if (pokemonForm) {
-			let pokemonIdString = `${pokeId}_${pokemonForm}${femaleSuffix}${costumeSuffix}${shinySuffix}`
-			if (availablePokemon.has(pokemonIdString)) return pokemonIdString
-			pokemonIdString = `${pokeId}_${pokemonForm}${costumeSuffix}${shinySuffix}`
-			if (availablePokemon.has(pokemonIdString)) return pokemonIdString
-			pokemonIdString = `${pokeId}_${pokemonForm}${femaleSuffix}${shinySuffix}`
-			if (availablePokemon.has(pokemonIdString)) return pokemonIdString
-			pokemonIdString = `${pokeId}_${pokemonForm}${shinySuffix}`
-			if (availablePokemon.has(pokemonIdString)) return pokemonIdString
-		}
-		const defaultForm = defaultForms[pokemonId]
-		if (defaultForm) {
-			let pokemonIdString = `${pokeId}_${defaultForm}${femaleSuffix}${costumeSuffix}${shinySuffix}`
-			if (availablePokemon.has(pokemonIdString)) return pokemonIdString
-			pokemonIdString = `${pokeId}_${defaultForm}${costumeSuffix}${shinySuffix}`
-			if (availablePokemon.has(pokemonIdString)) return pokemonIdString
-			pokemonIdString = `${pokeId}_${defaultForm}${femaleSuffix}${shinySuffix}`
-			if (availablePokemon.has(pokemonIdString)) return pokemonIdString
-			pokemonIdString = `${pokeId}_${defaultForm}${shinySuffix}`
-			if (availablePokemon.has(pokemonIdString)) return pokemonIdString
-		}
-		let pokemonIdString = `${pokeId}_00${femaleSuffix}${costumeSuffix}${shinySuffix}`
-		if (availablePokemon.has(pokemonIdString)) return pokemonIdString
-		pokemonIdString = `${pokeId}_00${costumeSuffix}${shinySuffix}`
-		if (availablePokemon.has(pokemonIdString)) return pokemonIdString
-		pokemonIdString = `${pokeId}_00${femaleSuffix}${shinySuffix}`
-		if (availablePokemon.has(pokemonIdString)) return pokemonIdString
-		pokemonIdString = `${pokeId}_00${shinySuffix}`
-		if (availablePokemon.has(pokemonIdString)) return pokemonIdString
-		return '000'	// substitute
 	}
 
 	getGeocoder() {
